Meeting: minor fix
