+++ lib/frontend/src/actions/circle.js
@@ -20,6 +20,11 @@ export const togglePurposeExpanded = purposeExpanded => ({
   purposeExpanded,
 })
 
+export const setCircle = circle => ({
+  type: 'SET_CIRCLE',
+  circle,
+})
+
 export const createCircle = (data, history) => async dispatch => {
   let response
   try {
@@ -205,8 +210,6 @@ export const fetchCircle = (
     .then(res => {
       // eslint-disable-next-line prefer-destructuring
       circle = res[0]
-      dispatch(setResults(circle, 'circle'))
-      users = res[2] ? res[2].results : state.users.results
       if (onlyCircle) {
         // from Issue detail or Create Role pages
         dispatch(setResults(circle, 'circle'))
@@ -215,6 +218,9 @@ export const fetchCircle = (
         }
         return circle
       }
+      dispatch(setCircle(circle))
+      users = res[2] ? res[2].results : state.users.results
+
       // add leadlind roles from circle children
       // to display user in attendees list and
       // for circle children leadlink 'election'
