+++ lib/frontend/src/actions/meetings.js
@@ -124,19 +124,6 @@ export const sortAttendees = presents => {
   return presentsList
 }
 
-export const sortProjects = milestones => {
-  milestones.sort(function(a, b) {
-    return `${a.repo.toLowerCase()}-${a.due_on}` >
-      `${b.repo.toLowerCase()}-${b.due_on}`
-      ? 1
-      : `${b.repo.toLowerCase()}-${b.due_on}` >
-        `${a.repo.toLowerCase()}-${a.due_on}`
-        ? -1
-        : 0
-  })
-  return milestones
-}
-
 export const updateReportsList = (event, history = null) => dispatch => {
   const params = {
     circle_id: event.target.form.circles.value
