+++ lib/frontend/src/reducer/gibolt.js
@@ -8,8 +8,14 @@ export const circle = (state = initial.circle, action) => {
         ...state,
         is_in_edition: !state.is_in_edition,
       }
+    case 'SET_CIRCLE':
+      return {
+        ...state,
+        results: action.circle,
+      }
     case 'TOGGLE_CIRCLE_ACCOUNT':
       return {
+        ...state,
         results: {
           ...state.results,
           accountabilities_expanded: !action.accountExpanded,
@@ -17,6 +23,7 @@ export const circle = (state = initial.circle, action) => {
       }
     case 'TOGGLE_CIRCLE_DOMAIN':
       return {
+        ...state,
         results: {
           ...state.results,
           domain_expanded: !action.domainExpanded,
@@ -24,6 +31,7 @@ export const circle = (state = initial.circle, action) => {
       }
     case 'TOGGLE_CIRCLE_PURPOSE':
       return {
+        ...state,
         results: {
           ...state.results,
           purpose_expanded: !action.purposeExpanded,
@@ -31,6 +39,7 @@ export const circle = (state = initial.circle, action) => {
       }
     case 'REMOVE_CIRCLE_MILESTONE':
       return {
+        ...state,
         results: {
           ...state.results,
           circle_milestones: action.circleMilestones,
