+++ lib/frontend/src/utils.js
@@ -327,6 +327,19 @@ export const sortUsers = users => {
   return users
 }
 
+export const sortProjects = milestones => {
+  milestones.sort(function(a, b) {
+    return `${a.repo.toLowerCase()}-${a.due_on}` >
+      `${b.repo.toLowerCase()}-${b.due_on}`
+      ? 1
+      : `${b.repo.toLowerCase()}-${b.due_on}` >
+        `${a.repo.toLowerCase()}-${a.due_on}`
+        ? -1
+        : 0
+  })
+  return milestones
+}
+
 export const roleTypes = [
   { value: 'leadlink', name: 'Premier lien' },
   { value: 'elected', name: 'Rôles élus' },
